working on login and register screens
