+++ src/app.jsx
@@ -13,8 +13,8 @@ import './index.scss';
 import Nav from './Components/Nav/Nav.jsx';
 import AddVocab from './screens/AddVocab.jsx';
 import SelectionScreen from './screens/SelectionScreen.jsx';
-import Login from './screens/Login.jsx';
-import Register from './screens/Register.jsx';
+import Login from './screens/Login/Login.jsx';
+import Register from './screens/Register/Register.jsx';
 import FirstStartup from './screens/FirstStartup.jsx';
 import TopNav from './Components/Nav/TopNav.jsx';
