+++ src/screens/Register/Register.scss
@@ -1,46 +1,62 @@
+@import "../../colors.scss";
+
 .register {
-    width: "100%";
-    height: "100vh";
+    width: 100%;
+    height: 100vh;
     display: flex;
     justify-content: center;
-    background: #F1F3FA;
-    background-image: linear-gradient(to top right, #2e5695, #F7D0FB);
+    background: #f1f3fa;
+    background-image: linear-gradient(to top right, #2e5695, #f7d0fb);
 
     .register-form {
-        width: "25%";
-        height: "70%";
+        max-width: 450px;
+        max-height: 550px;
+
         margin: auto;
+        border-radius: 10px;
+
         display: flex;
         justify-content: space-around;
         flex-direction: column;
-        padding: 40px;
+
+        padding: 40px 100px;
+        background: $color-white;
         box-shadow: -1px 3px 5px#b4bed6;
-        background
 
         .register-form-header {
-            width: "100%";
-            height: "33%";
+            width: 100%;
+            height: 33%;
+
             .register-form-header-logo {
-                height: "70%";
-                margin: auto;
+                width: 100px;
+                height: auto;
+                margin-bottom: 10px;
+            }
+
+            .register-form-header-heading {
+                text-transform: uppercase;
             }
         }
 
         .register-form-input {
-            width: "100%";
+            width: 100%;
+            min-width: 200px;
+            margin: 30px auto;
+            display: flex;
+            flex-direction: column;
+            justify-content: space-around;
 
             .register-form-input-field {
-                width: "80%";
+                width: 80%;
                 height: 35px;
                 background: transparent;
                 outline: none;
                 border: none;
-                border-bottom: 1px solid #8790A3;
+                border-bottom: 1px solid #8790a3;
                 text-align: left;
-                margin: 20px 0;
-
+                margin: 10px 0;
             }
-            
+
             .register-form-warning__invisible {
                 display: none;
             }
@@ -52,34 +68,17 @@
         }
 
         .register-form-submit {
-            .register-form-submit-Btn {
-                width: "80%";
-                height: 45px;
-                background: #727CF5;
-                outline: none;
-                border: none;
-                border-radius: 15px;
-                color: #FFFFFF;
-                font-size: 20px;
-                cursor: pointer;
-                margin: 20px 0 0 0;
-
-                &:hover{
-                    background: #6771df;
-                }
-            }
-
             .register-form-submit-register {
+                display: flex;
                 font-size: 14px;
+                margin-top: 12px;
 
                 .register-form-submit-register-link {
                     color: blue;
                     cursor: pointer;
+                    margin-left: 5px;
                 }
-            } 
+            }
         }
-
-        
-
     }
 }
